// src/generated/sharevox/api/models/UserDictWord.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 辞書のコンパイルに使われる情報
 */
export type UserDictWord = {
    surface: string;
    priority: number;
    context_id?: number;
    part_of_speech: string;
    part_of_speech_detail_1: string;
    part_of_speech_detail_2: string;
    part_of_speech_detail_3: string;
    inflectional_type: string;
    inflectional_form: string;
    stem: string;
    yomi: string;
    pronunciation: string;
    accent_type: number;
    mora_count?: number;
    accent_associative_rule: string;
};